Completes sqlite database config setup
